+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Update dependencies
+- Polkadot Vault `signBytes`
 
 ## 0.3.0 2025-11-10